Update record approval logic
A recrod needs at least two "valid votes" to get into the approved state. If a record was reviewed by at least 3 persons and and there are not at least two valid votes, the record is marked as invalid.
This refactoring does not change this logic, but allows records to be moved to the approved state already after two reviews (if both approve the record).
This optimizes the review-time, as the third reviewer would have no effect.
Fixes #224