fix: prevent horizontal scrollbar from flickering during swipe
When reviewing sentences the TinderCard animation will make the horizontal (and
sometimes the vertical scrollbar) appear. This is quite jarring to me.

The TinderCard library actually recommend using `overflow: hidden`, but we need
to allow vertical scrolling.

As far as I could see no of the other pages need horizontal scrolling though -
so we can at least fix one of the cases.