feat: add UI language selector to profile (fixes #533) (#563)
This adds a UI language selector to the profile. It will show all the languages for now. We can't reliably use translated.json as it really does not give much information on how much of Sentence Collector is actually translated. Therefore I went with that approach. However, if a locale is not in translated.json we show a warning with a link to Pontoon to help out with translation.